it looks completed

+++ ukol08_Burjanova/src/main/java/cz/czechitas/java/database/SpousteciTrida.java
@@ -14,7 +14,7 @@ import java.util.Scanner;
 
 public class SpousteciTrida {
 
-    public static void main(String[] args) throws SQLException {
+    public static void main(String[] args) throws SQLException, InterruptedException {
         MariaDbDataSource konfiguraceDatabaze = new MariaDbDataSource();
         konfiguraceDatabaze.setUrl("jdbc:mysql://localhost:3306/DailyPlanetMartina");
         konfiguraceDatabaze.setUserName("student");
@@ -23,7 +23,7 @@ public class SpousteciTrida {
         JdbcTemplate odesilacDotazu = new JdbcTemplate(konfiguraceDatabaze);
 
         ResultSetExtractor<Collection<Clanek>> prevodnikClankuSAutory = new PrevodnikClankuSAutory();
-        Collection<Clanek> clanky3 = odesilacDotazu.query("" +
+        Collection<Clanek> clanky1 = odesilacDotazu.query("" +
             "select clanky.idClanku as idClanku, clanky.nazev as nazev, clanky.datum as datum, " +
             "       zamestnanci.idAutor as autorId, zamestnanci.jmeno as jmeno, zamestnanci.bydliste as bydliste, zamestnanci.plat, zamestnanci.datumNastupu " +
             "  from clanky join zamestnanci on clanky.idAutor=zamestnanci.idAutor", prevodnikClankuSAutory);
@@ -31,7 +31,7 @@ public class SpousteciTrida {
         System.out.println("*************************************************************************");
         System.out.println("Seznam všech článků Daily Planet:");
         System.out.println();
-        for (Clanek clanek : clanky3) {
+        for (Clanek clanek : clanky1) {
             System.out.println("Autor: " + clanek.getAutor().getJmeno() + ", název: " + clanek.getNazev());
         }
 
@@ -39,8 +39,8 @@ public class SpousteciTrida {
         System.out.println("*************************************************************************");
         System.out.println("Seznam všech článků Daily Planet s datem nástupu redaktorů:");
         System.out.println();
-        for (Clanek clanek : clanky3) {
-            System.out.println("Název článku: " + clanek.getNazev() + " od " + clanek.getAutor().getJmeno() + " v Dialy Planet od: " + clanek.getAutor().getDatumNastupu() );
+        for (Clanek clanek : clanky1) {
+            System.out.println("Název článku: " + clanek.getNazev() + " od " + clanek.getAutor().getJmeno() + " v Dialy Planet od: " + clanek.getAutor().getDatumNastupu());
         }
 
         //získej počet článků:
@@ -49,7 +49,8 @@ public class SpousteciTrida {
         System.out.println("*************************************************************************");
         System.out.println("V databázi je " + pocetClanku + " článků");
 
-        RowMapper<Zamestnanec> prevodnikZamci;System.out.println();
+        RowMapper<Zamestnanec> prevodnikZamci;
+        System.out.println();
         prevodnikZamci = BeanPropertyRowMapper.newInstance(Zamestnanec.class);
         List<Zamestnanec> zamci = odesilacDotazu.query("select * from zamestnanci", prevodnikZamci);
         //vypíše seznam všech zaměstnanců
@@ -64,22 +65,26 @@ public class SpousteciTrida {
         //vypsat články autora podle id:
         System.out.println();
         System.out.println("*************************************************************************");
-        System.out.println("Zadejte ID zaměstnance a stiskněte enter, vypíšeme její/jeho články:");
+        System.out.println("Zadejte ID zaměstnance a stiskněte enter, vypíšeme její/jeho články, zaměstnanci jsou 4:");
         System.out.println();
         Scanner sc = new Scanner(System.in);
         Long kodZamestnanec = sc.nextLong();
 
-        String hledany_autor = odesilacDotazu.queryForObject("select distinct Autor from clanky where idAutor=?", String.class, kodZamestnanec);
 
-        System.out.println();
-        System.out.println("*************************************************************************");
-        System.out.println("Seznam článků autora " + hledany_autor + ":");
-        System.out.println();
+        if (kodZamestnanec < 5) {
+            String hledany_autor = odesilacDotazu.queryForObject("select distinct Autor from clanky where idAutor=?", String.class, kodZamestnanec);
+            System.out.println();
+            System.out.println("*************************************************************************");
+            System.out.println("Seznam článků autora " + hledany_autor + ":");
+            System.out.println();
 
-        for (Clanek c : clanky3) {
-            if (c.getAutor().getId() == kodZamestnanec) {
-                System.out.println(c.clankysPodleAutora());
+            for (Clanek c : clanky1) {
+                if (c.getAutor().getId() == kodZamestnanec) {
+                    System.out.println(c.clankysPodleAutora());
+                }
             }
+        } else {
+            System.out.println("Nezadali jste správný kód zaměstnance, program pokračuje");
         }
 
 
@@ -100,17 +105,17 @@ public class SpousteciTrida {
         System.out.println("Zkuste hádat, ve který den v týdnu v únoru 2019 se Daily Planet nejlépe prodával, napište pondělí, úterý, " +
             "středa, čtvrtek nebo pátek a stiskněte enter");
         System.out.println();
-        Scanner sc2 = new Scanner(System.in);
-        String tipDenTydnu = sc2.next();
+        String tipDenVTydnu = sc.next();
 
         //   průměrný prodaný náklad podle dne v týdnu:
         RowMapper<Naklad> prevodnikNaklad;
         prevodnikNaklad = BeanPropertyRowMapper.newInstance(Naklad.class);
         String topDenVTydnu = odesilacDotazu.queryForObject("SELECT denVTydnu FROM prodanyNaklad GROUP BY denVTydnu order by prodanyNaklad desc limit 1;", String.class);
-        if (topDenVTydnu == tipDenTydnu) {
+
+        if (topDenVTydnu.equals(tipDenVTydnu)) {
             System.out.println("Správně, " + topDenVTydnu + " je den v týdnu, kdy se prodalo nejvíc novin Daily Planet");
         } else {
-            System.out.println("Ne, " + tipDenTydnu + " není den, kdy se prodalo nejvíc novin Daily Planet.");
+            System.out.println("Ne, " + tipDenVTydnu + " není den, kdy se prodalo nejvíc novin Daily Planet.");
         }
 
         //  vypiš prodaný náklad podle dnů v týdnu:
@@ -140,12 +145,38 @@ public class SpousteciTrida {
         System.out.println("*************************************************************************");
         System.out.println("Napište číslo článku, který chcete smazat a stiskněte enter, článků je celkem " + pocetClanku);
         System.out.println();
-        Scanner sc3 = new Scanner(System.in);
-        String clanekNaSmazani = sc3.next();
+        Long clanekNaSmazani = sc.nextLong();
 
-        odesilacDotazu.update("delete from clanky where idClanku=?", clanekNaSmazani);
+        for (Clanek d : clanky1) {
+            if (d.getId() == clanekNaSmazani) {
+                System.out.println(d.clankysPodleAutora());
+                odesilacDotazu.update("delete from clanky where idClanku=?", clanekNaSmazani);
 
+                System.out.println("Smazali jste celou databázi, ajajaj");
+                Thread.sleep(2000);
+
+                System.out.println("to byl vtip");
+
+                Collection<Clanek> clanky2 = odesilacDotazu.query("" +
+                    "select clanky.idClanku as idClanku, clanky.nazev as nazev, clanky.datum as datum, " +
+                    "       zamestnanci.idAutor as autorId, zamestnanci.jmeno as jmeno, zamestnanci.bydliste as bydliste, zamestnanci.plat, zamestnanci.datumNastupu " +
+                    "  from clanky join zamestnanci on clanky.idAutor=zamestnanci.idAutor", prevodnikClankuSAutory);
+
+                System.out.println();
+
+                System.out.println();
+                for (Clanek clanek : clanky2) {
+                    System.out.println("Autor: " + clanek.getAutor().getJmeno() + ", název: " + clanek.getNazev());
+                }
+
+                Long pocetClanku2 = odesilacDotazu.queryForObject("select count (*) from clanky", Long.class);
+                System.out.println();
+                System.out.println("V databázi už je jen " + pocetClanku2 + " článků");
+
+                return;
+            }
+        }
+        System.out.println("Zadaný článek v databázi neexistuje");
 
     }
-
 }
